+++ examples/ErrorExample.hs
@@ -24,7 +24,8 @@ templateVars = M.fromList
 
 main :: IO ()
 main = do
-  let pandocWriterConfig = K.PandocWriterConfig (Just "pandoc-templates/minWithVega-pandoc.html")  templateVars K.mindocOptionsF
+  let template =  K.FromIncludedTemplateDir "pandoc-bootstrap-KH.html"
+  pandocWriterConfig <- K.mkPandocWriterConfig template templateVars K.mindocOptionsF
   resSimpleE <- K.knitHtml (Just "SimpleExample.Main") K.logAll pandocWriterConfig makeDocWithKnitError
   case resSimpleE of
     Right htmlAsText ->